requirements.txt
# Vercel Python requirements - mirrors backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai>=1.50.0
supabase>=2.10.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
httpx>=0.28.0
pdfplumber>=0.11.0
aiofiles>=24.1.0
mangum>=0.17.0
